+++ src/components/HeatmapHeader/HeatmapHeader.tsx
@@ -1,9 +1,31 @@
 import { useTranslation } from "react-i18next";
 import { useHeatmapContext } from "src/context/heatmap/heatmap.context";
+import { View } from "src/types";
+
+function StatisticsIcon() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="18"
+      height="18"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      className="lucide lucide-chart-line"
+    >
+      <path d="M3 3v16a2 2 0 0 0 2 2h16" />
+      <path d="m19 9-5 5-4-4-3 3" />
+    </svg>
+  );
+}
 
 export function HeatmapHeader() {
   const { t } = useTranslation();
-  const { currentYear, setCurrentYear, trackerData } = useHeatmapContext();
+  const { currentYear, setCurrentYear, trackerData, setView } =
+    useHeatmapContext();
 
   function onArrowBackClick() {
     setCurrentYear(currentYear - 1);
@@ -38,24 +60,17 @@ export function HeatmapHeader() {
         <div className="heatmap-tracker-header__title">
           {trackerData?.heatmapTitle ?? ""}
         </div>
-        <div className="heatmap-tracker-header__settings">
-          {/* <button onClick={() => setView(View.HeatmapTrackerSettings)}>
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              width="24"
-              height="24"
-              viewBox="0 0 24 24"
-              fill="none"
-              stroke="currentColor"
-              strokeWidth="2"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              className="lucide lucide-settings"
-            >
-              <path d="M12.22 2h-.44a2 2 0 0 0-2 2v.18a2 2 0 0 1-1 1.73l-.43.25a2 2 0 0 1-2 0l-.15-.08a2 2 0 0 0-2.73.73l-.22.38a2 2 0 0 0 .73 2.73l.15.1a2 2 0 0 1 1 1.72v.51a2 2 0 0 1-1 1.74l-.15.09a2 2 0 0 0-.73 2.73l.22.38a2 2 0 0 0 2.73.73l.15-.08a2 2 0 0 1 2 0l.43.25a2 2 0 0 1 1 1.73V20a2 2 0 0 0 2 2h.44a2 2 0 0 0 2-2v-.18a2 2 0 0 1 1-1.73l.43-.25a2 2 0 0 1 2 0l.15.08a2 2 0 0 0 2.73-.73l.22-.39a2 2 0 0 0-.73-2.73l-.15-.08a2 2 0 0 1-1-1.74v-.5a2 2 0 0 1 1-1.74l.15-.09a2 2 0 0 0 .73-2.73l-.22-.38a2 2 0 0 0-2.73-.73l-.15.08a2 2 0 0 1-2 0l-.43-.25a2 2 0 0 1-1-1.73V4a2 2 0 0 0-2-2z" />
-              <circle cx="12" cy="12" r="3" />
-            </svg>
-          </button> */}
+        <div className="heatmap-tracker-header__statistics">
+          <button
+            onClick={() => setView(View.HeatmapTrackerStatistics)}
+            className="clickable-icon"
+            aria-label="Statistics"
+          >
+            <>
+              <StatisticsIcon />
+              <span>(beta)</span>
+            </>
+          </button>
         </div>
       </div>
       {trackerData?.heatmapSubtitle ? (